requirements.txt
streamlit>=1.28.0
openai>=1.12.0,<2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0,<1.0.0

